requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
ariadne-codegen==0.14.0
asgiref==3.10.0
astroid==3.3.11
autoflake==2.3.1
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dill==0.4.0
distro==1.9.0
Django==5.2.7
django-cors-headers==4.9.0
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
expiringdict==1.2.2
filelock==3.20.0
fsspec==2025.10.0
googleapis-common-protos==1.72.0
graphql-core==3.2.7
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.6
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.3.0
isort==6.1.0
jiter==0.12.0
launchdarkly-eventsource==1.5.0
launchdarkly-observability==0.1.1
launchdarkly-server-sdk==9.13.1
launchdarkly-server-sdk-otel==1.2.0
mccabe==0.7.0
mypy_extensions==1.1.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-aiohttp-client==0.55b1
opentelemetry-instrumentation-aiohttp-server==0.55b1
opentelemetry-instrumentation-anthropic==0.40.14
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-bedrock==0.40.14
opentelemetry-instrumentation-boto==0.55b1
opentelemetry-instrumentation-boto3sqs==0.55b1
opentelemetry-instrumentation-celery==0.55b1
opentelemetry-instrumentation-chromadb==0.40.14
opentelemetry-instrumentation-cohere==0.40.14
opentelemetry-instrumentation-dbapi==0.55b1
opentelemetry-instrumentation-django==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-instrumentation-flask==0.55b1
opentelemetry-instrumentation-grpc==0.55b1
opentelemetry-instrumentation-haystack==0.40.14
opentelemetry-instrumentation-httpx==0.55b1
opentelemetry-instrumentation-jinja2==0.55b1
opentelemetry-instrumentation-langchain==0.40.14
opentelemetry-instrumentation-llamaindex==0.40.14
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-openai==0.40.14
opentelemetry-instrumentation-pinecone==0.40.14
opentelemetry-instrumentation-psycopg==0.55b1
opentelemetry-instrumentation-psycopg2==0.55b1
opentelemetry-instrumentation-qdrant==0.40.14
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.40.14
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-sqlite3==0.55b1
opentelemetry-instrumentation-starlette==0.55b1
opentelemetry-instrumentation-system-metrics==0.55b1
opentelemetry-instrumentation-transformers==0.40.14
opentelemetry-instrumentation-urllib==0.55b1
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.40.14
opentelemetry-instrumentation-watsonx==0.40.14
opentelemetry-instrumentation-weaviate==0.40.14
opentelemetry-instrumentation-wsgi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.55b1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
protobuf==5.29.5
psutil==7.1.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyRFC3339==2.1.0
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.1
pytokens==0.3.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
semver==3.0.4
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
zipp==3.23.0
